+++ server/main.go
@@ -5,27 +5,13 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/bufbuild/connect-go"
 	"golang.org/x/exp/slog"
 
-	qrurlv1 "github.com/emahiro/qrurl/server/gen/proto/qrurl/v1"
 	"github.com/emahiro/qrurl/server/gen/proto/qrurl/v1/qrurlv1connect"
+	"github.com/emahiro/qrurl/server/service.go"
 )
 
 const addr = ":8080"
-
-type QrUrlService struct{}
-
-func (s *QrUrlService) PostCode(
-	ctx context.Context,
-	req *connect.Request[qrurlv1.PostCodeRequest],
-) (resp *connect.Response[qrurlv1.PostCodeResponse], err error) {
-	qrurlResp := &qrurlv1.PostCodeResponse{
-		Url: "test",
-	}
-	resp = connect.NewResponse(qrurlResp)
-	return resp, nil
-}
 
 func main() {
 	ctx := context.Background()
@@ -33,7 +19,7 @@ func main() {
 	defer cancel()
 
 	mux := http.NewServeMux()
-	mux.Handle(qrurlv1connect.NewQrUrlServiceHandler(&QrUrlService{}))
+	mux.Handle(qrurlv1connect.NewQrUrlServiceHandler(&service.QrUrlService{}))
 	server := &http.Server{
 		Addr:    addr,
 		Handler: mux,
